R script to subdivide files for x-tab.
R script works around memory limitations by splitting files.
Segments are input to xtab() and the result written to disk.
Files are erased from memory with only the column means kept
in live memory. This collection of means is used to limit
row numbers (variables) when split files are recombined for
classification testing.